Remove Newtonsoft.Json dependency Separate Msmq Medium in Library
