+++ .github/workflows/dotnet5.yml
@@ -37,6 +37,13 @@ jobs:
         PACKAGE_NAME: Ariane5.Azure
         NUGET_SOURCE: https://api.nuget.org
         NUGET_KEY: ${{secrets.nuget}}
+    - name: Publish Ariane5.Azure to nuget
+      uses: rohith/publish-nuget@v2
+      with:
+        PROJECT_FILE_PATH: ${{github.workspace}}\src\Ariane5.Azure\Ariane5.Msmq.csproj
+        PACKAGE_NAME: Ariane5.Msmq
+        NUGET_SOURCE: https://api.nuget.org
+        NUGET_KEY: ${{secrets.nuget}}
     - name: Publish Ariane5 Standard to nuget
       uses: rohith/publish-nuget@v2
       with:
