+++ .gitignore
@@ -216,3 +216,4 @@ ModelManifest.xml
 /.vs/Ariane5/config
 /.vs/Ariane5/v16/TestStore/OutputMessages
 /.vs/Ariane5/DesignTimeBuild
+/.vs/Ariane5
