+++ Ariane5.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30011.22
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{1E435FBE-FD77-45EF-865D-E3B6C3CD60E8}"
 	ProjectSection(SolutionItems) = preProject
@@ -10,8 +10,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		.github\workflows\dotnet5.yml = .github\workflows\dotnet5.yml
 		README.md = README.md
 	EndProjectSection
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "dotnet5", "dotnet5", "{6A8001D9-63F6-4264-B140-7B4C54B60A75}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Ariane5", "src\Ariane5\Ariane5.csproj", "{E2B9436A-801F-4594-BC5A-B6F8B26D8056}"
 EndProject
@@ -33,11 +31,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Ariane5.AzureTopicClientWri
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Ariane5.AzureTopicClientReaders", "src\Azure5.AzureTopicClientReaders\Ariane5.AzureTopicClientReaders.csproj", "{A479AD31-32F7-4C9D-8D54-6B2DF49518A7}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "dotnet.standard", "dotnet.standard", "{4B6308BA-AC2F-4863-AEC7-6F9EA29D497C}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ArianeStandard", "src\ArianeStandard\ArianeStandard.csproj", "{1B49D89E-9948-44E1-8E1A-9ED8144B7169}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ArianeStandard.Azure", "src\ArianeStandard.Azure\ArianeStandard.Azure.csproj", "{FCA8933A-2E1F-4820-81BA-0AB1ABF386D8}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Ariane5.Msmq", "src\Ariane5.Msmq\Ariane5.Msmq.csproj", "{37F32BE3-6480-48C6-A347-ADF766147411}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -81,31 +75,21 @@ Global
 		{A479AD31-32F7-4C9D-8D54-6B2DF49518A7}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A479AD31-32F7-4C9D-8D54-6B2DF49518A7}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A479AD31-32F7-4C9D-8D54-6B2DF49518A7}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1B49D89E-9948-44E1-8E1A-9ED8144B7169}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1B49D89E-9948-44E1-8E1A-9ED8144B7169}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1B49D89E-9948-44E1-8E1A-9ED8144B7169}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1B49D89E-9948-44E1-8E1A-9ED8144B7169}.Release|Any CPU.Build.0 = Release|Any CPU
-		{FCA8933A-2E1F-4820-81BA-0AB1ABF386D8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{FCA8933A-2E1F-4820-81BA-0AB1ABF386D8}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{FCA8933A-2E1F-4820-81BA-0AB1ABF386D8}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{FCA8933A-2E1F-4820-81BA-0AB1ABF386D8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{37F32BE3-6480-48C6-A347-ADF766147411}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{37F32BE3-6480-48C6-A347-ADF766147411}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{37F32BE3-6480-48C6-A347-ADF766147411}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{37F32BE3-6480-48C6-A347-ADF766147411}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{E2B9436A-801F-4594-BC5A-B6F8B26D8056} = {6A8001D9-63F6-4264-B140-7B4C54B60A75}
-		{C1B38AE6-C1F7-4B32-9324-FD663BFF33E4} = {6A8001D9-63F6-4264-B140-7B4C54B60A75}
-		{73CF29A5-5E70-4AD9-9B3C-A89D31FD312C} = {6A8001D9-63F6-4264-B140-7B4C54B60A75}
 		{E6A9B681-C3E6-4879-B845-514C34364E74} = {F78FC700-35EF-40DC-9954-0648EDC87994}
 		{CE65098A-546A-4932-A006-EEF83E08FF4E} = {F78FC700-35EF-40DC-9954-0648EDC87994}
 		{6B8C7FE6-6766-43DC-805A-0C685B4A2D4F} = {F78FC700-35EF-40DC-9954-0648EDC87994}
 		{DBEDB377-4E3F-4BE6-9371-FCC977F8EF27} = {F78FC700-35EF-40DC-9954-0648EDC87994}
-		{F78FC700-35EF-40DC-9954-0648EDC87994} = {6A8001D9-63F6-4264-B140-7B4C54B60A75}
 		{69FA854C-448B-402A-87D8-8BE1CFDAF5E6} = {F78FC700-35EF-40DC-9954-0648EDC87994}
 		{A479AD31-32F7-4C9D-8D54-6B2DF49518A7} = {F78FC700-35EF-40DC-9954-0648EDC87994}
-		{1B49D89E-9948-44E1-8E1A-9ED8144B7169} = {4B6308BA-AC2F-4863-AEC7-6F9EA29D497C}
-		{FCA8933A-2E1F-4820-81BA-0AB1ABF386D8} = {4B6308BA-AC2F-4863-AEC7-6F9EA29D497C}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {82518F7F-349E-4013-92A6-ABB7F5B87E24}
