+++ src/Ariane5.Azure/Ariane5.Azure.csproj
@@ -3,8 +3,12 @@
   <PropertyGroup>
     <TargetFramework>net5.0</TargetFramework>
     <RootNamespace>Ariane</RootNamespace>
-    <Version>5.1.10</Version>
+    <Version>5.2.11</Version>
   </PropertyGroup>
+
+  <ItemGroup>
+    <Compile Remove="ReceivedMessage.cs" />
+  </ItemGroup>
 
   <ItemGroup>
     <PackageReference Include="Azure.Messaging.ServiceBus" Version="7.1.0" />
