+++ src/Ariane5.Azure/QueueProvider/AzureMessageQueue.cs
@@ -10,8 +10,6 @@ using Azure.Messaging.ServiceBus;
 
 using Microsoft.Extensions.Logging;
 
-using Newtonsoft.Json;
-
 namespace Ariane.QueueProviders
 {
 	public class AzureMessageQueue : IMessageQueue, IAsyncDisposable, IDisposable
@@ -84,7 +82,7 @@ namespace Ariane.QueueProviders
 		{
 			if (m_BinaryMessage != null)
             {
-				var receiveMessage = JsonConvert.DeserializeObject<T>(m_BinaryMessage.ToString());
+				var receiveMessage = System.Text.Json.JsonSerializer.Deserialize<T>(m_BinaryMessage.ToString());
 				return receiveMessage;
 			}
 			return default(T);
@@ -97,7 +95,7 @@ namespace Ariane.QueueProviders
 			if (receivedMessage != null
 				&& receivedMessage.Body != null)
             {
-				var receiveMessage = JsonConvert.DeserializeObject<T>(receivedMessage.Body.ToString());
+				var receiveMessage = System.Text.Json.JsonSerializer.Deserialize<T>(receivedMessage.Body.ToString());
 				result = receiveMessage; 
 			}
 			return result;
@@ -125,7 +123,7 @@ namespace Ariane.QueueProviders
 				string data = null;
 				try
 				{
-					data = JsonConvert.SerializeObject(message.Body);
+					data = System.Text.Json.JsonSerializer.Serialize(message.Body);
 					busMessage = new ServiceBusMessage(System.Text.Encoding.UTF8.GetBytes(data));
 					busMessage.Subject = message.Label;
 					if (message.ScheduledEnqueueTimeUtc.HasValue)
@@ -165,7 +163,7 @@ namespace Ariane.QueueProviders
 				var batch = await m_ServiceBusSender.CreateMessageBatchAsync(new CancellationToken());
 				foreach (var message in messages)
 				{
-					var data = JsonConvert.SerializeObject(message.Body);
+					var data = System.Text.Json.JsonSerializer.Serialize(message.Body);
 					var busMessage = new ServiceBusMessage(System.Text.Encoding.UTF8.GetBytes(data));
 					busMessage.Subject = message.Label;
 					if (message.ScheduledEnqueueTimeUtc.HasValue)
