+++ src/Ariane5.Msmq/Mediums/MSMQMedium.cs
