+++ src/Ariane5.Tests/Ariane5.Tests.csproj
@@ -37,6 +37,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\Ariane5.Azure\Ariane5.Azure.csproj" />
+    <ProjectReference Include="..\Ariane5.Msmq\Ariane5.Msmq.csproj" />
     <ProjectReference Include="..\Ariane5\Ariane5.csproj" />
   </ItemGroup>
