+++ src/Ariane5/RegisterExtensions.cs
@@ -73,70 +73,6 @@ namespace Ariane
 		}
 
 		/// <summary>
-		/// Add MSMQ Reader
-		/// </summary>
-		/// <param name="register"></param>
-		/// <param name="queueName"></param>
-		/// <param name="typeReader"></param>
-		/// <returns></returns>
-		public static IRegister AddMSMQReader(this IRegister register, string queueName, Type typeReader)
-		{
-			if (queueName == null || register == null)
-			{
-				throw new ArgumentNullException();
-			}
-
-			var queueSetting = new QueueSetting()
-			{
-				Name = queueName,
-				TypeReader = typeReader,
-				TypeMedium = typeof(MSMQMedium),
-			};
-			register.AddQueue(queueSetting);
-			return register;
-		}
-
-		public static IRegister AddMSMQReader<T>(this IRegister register, string queueName)
-		{
-			if (queueName == null || register == null)
-			{
-				throw new ArgumentNullException();
-			}
-
-			var queueSetting = new QueueSetting()
-			{
-				Name = queueName,
-				TypeReader = typeof(T),
-				TypeMedium = typeof(MSMQMedium),
-			};
-			register.AddQueue(queueSetting);
-			return register;
-		}
-
-
-		/// <summary>
-		/// Add MSMQ Writer
-		/// </summary>
-		/// <param name="register"></param>
-		/// <param name="queueName"></param>
-		/// <returns></returns>
-		public static IRegister AddMSMQWriter(this IRegister register, string queueName)
-		{
-			if (queueName == null || register == null)
-			{
-				throw new ArgumentNullException();
-			}
-
-			var queueSetting = new QueueSetting()
-			{
-				Name = queueName,
-				TypeMedium = typeof(MSMQMedium),
-			};
-			register.AddQueue(queueSetting);
-			return register;
-		}
-
-		/// <summary>
 		/// Add File Reader
 		/// </summary>
 		/// <param name="register"></param>
